+++ lab_08_04/src/userio.c
@@ -2,8 +2,8 @@
 #include <inttypes.h>
 #include <stdio.h>
 
-static exit_t input_int(int *value);
-static exit_t print_int(int value, char end);
+static exit_t input_int(int64_t *value);
+static exit_t print_int(int64_t value, char end);
 
 exit_t input_matrix(matrix_t *matrix)
 {
@@ -17,12 +17,12 @@ exit_t input_matrix(matrix_t *matrix)
 
     if (exit_code == exit_success)
     {
-        int **pp_end = matrix->element + matrix->rows;
+        int64_t **pp_end = matrix->element + matrix->rows;
 
-        for (int **pp = matrix->element; exit_code == exit_success && pp < pp_end; pp++)
+        for (int64_t **pp = matrix->element; exit_code == exit_success && pp < pp_end; pp++)
         {
-            int *p_end = *pp + matrix->cols;
-            for (int *p = *pp; exit_code == exit_success && p < p_end; p++)
+            int64_t *p_end = *pp + matrix->cols;
+            for (int64_t *p = *pp; exit_code == exit_success && p < p_end; p++)
                 exit_code = input_int(p);
         }
 
@@ -57,13 +57,13 @@ exit_t print_matrix(const matrix_t *matrix)
 {
     exit_t exit_code = exit_success;
 
-    int **pp_end = matrix->element + matrix->rows;
+    int64_t **pp_end = matrix->element + matrix->rows;
 
-    for (int **pp = matrix->element; exit_code == exit_success && pp < pp_end; pp++)
+    for (int64_t **pp = matrix->element; exit_code == exit_success && pp < pp_end; pp++)
     {
-        int *p_end = *pp + matrix->cols;
+        int64_t *p_end = *pp + matrix->cols;
 
-        for (int *p = *pp; exit_code == exit_success && p < p_end; p++)
+        for (int64_t *p = *pp; exit_code == exit_success && p < p_end; p++)
         {
             char ch_end = (p_end - p == 1) ? '\n' : ' ';
             exit_code = print_int(*p, ch_end);
@@ -73,11 +73,11 @@ exit_t print_matrix(const matrix_t *matrix)
     return exit_code;
 }
 
-exit_t input_int(int *value)
+exit_t input_int(int64_t *value)
 {
     exit_t exit_code = exit_success;
 
-    int res = scanf("%d", value);
+    int res = scanf("%ld", value);
 
     if (res == 0)
         exit_code = exit_invalid_input;
@@ -87,11 +87,11 @@ exit_t input_int(int *value)
     return exit_code;
 }
 
-exit_t print_int(int value, char end)
+exit_t print_int(int64_t value, char end)
 {
     exit_t exit_code = exit_success;
 
-    if (printf("%d%c", value, end) <= 1)
+    if (printf("%ld%c", value, end) <= 1)
         exit_code = exit_output_failure;
 
     return exit_code;
